Refactor views.py to handle empty events from LINE DEVELOPERS
This change updates the views.py to handle cases where events are empty. Particularly when running validations from the LINE DEVELOPERS screen, an empty list of events is returned which could cause an IndexError in the previous implementation. The HttpResponse message has also been updated to "`callback` returned 200" for additional clarity.